pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool]
[tool.poetry]
authors = ["timelessnesses <mooping3roblox@gmail.com>"]
description = ""
name = "application-manager-bot"
version = "0.1.0"
[tool.poetry.dependencies]
python = "^3.8"
aiofiles = "^0.8.0"
Flask = "^2.0.2"
python-dotenv = "^0.19.2"
"discord.py" = {git = "https://github.com/Rapptz/discord.py"}
uvloop = {version = "^0.16.0", platform = "linux"}
orjson = "^3.6.7"
requests = "^2.27.1"
isort = "^5.10.1"
[tool.poetry.dependencies.aiohttp]
extras = ["speedups"]
version = "<3.8.0"
[tool.poetry.dev-dependencies]
black = "^21.12b0"
isort = "^5.10.1"
